backend/requirements.txt
# FastAPI & Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Message Queue & Database
kafka-python==2.0.2
clickhouse-driver==0.2.6
qdrant-client==1.7.3

# Log Analysis & Anomaly Detection
drain3==0.9.11
pyod==1.1.3

# Data Processing
numpy==1.26.3
pandas==2.1.4

# AI & LLM
langchain==0.2.16  # Pydantic v2 지원 버전
langchain-core==0.2.38
langchain-openai==0.1.25
langchain-google-genai==1.0.10
langgraph==0.2.28  # 최신 안정 버전
langsmith==0.1.129  # 명시적 버전 고정
openai==1.51.2
google-generativeai==0.8.3

# Embeddings
sentence-transformers==2.5.1
torch==2.2.0

# HTTP & Utils
requests==2.31.0
httpx==0.26.0
python-dotenv==1.0.0

# System Monitoring
psutil==5.9.8
